Moved to package Page indexing task files
